WQP-1592 (Drop "unlogged" keyword from all WQP tables)    Removed unlogged from hdatumToSrid.sql    Update d to version v2 for the create changeLog
The approach taken here is that, since hdatum_to_srid is a regular table
with no partitions (small, 16K currently on test), the setting
of the table to logged can be done separately using a single
sql command existing databases. Going forward when this is
used to create the table, the table will be in the desired state.
No need to maintain a separate lquidbase script.